fix(function1): 디자인 변경
