+++ app/src/main/res/values/colors.xml
@@ -17,4 +17,7 @@
     <color name="home_gray_color">#A6A6A6</color>
     <color name="home_arrow_gray_color">#D4D4D4</color>
     <color name="home_not_food">#E31F1F</color>
+
+<!--    custom color-->
+    <color name="c_656565">#656565</color>
 </resources>
